feat(style): adjust layout
Keep the header and footer at the top and bottom respectively.
Add padding to header, footerm and buttons.
Links in header and footer are pushed to either side.
Add spacing between links in header and footer and between the logo,
input, and buttons.
Logo, input, and buttons are vertically and horizontally centered.